+++ packages/orq-rc/RELEASES.md
@@ -2389,3 +2389,13 @@ Based on:
 - [typescript v3.9.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v3.9.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.8 - packages/orq-rc
+
+## 2025-06-30 06:36:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.571.2 (2.640.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.9.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v3.9.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.9 - packages/orq-rc